refactored socketMiddleware
